requirements.txt
markdown
restframework
django-filter
dj-rest-auth